# wasm: remove opa_number_float