--- a/wasm/src/conversions.c
+++ b/wasm/src/conversions.c
@@ -20,13 +20,13 @@ opa_value *opa_to_number(opa_value *v)
     {
         opa_string_t *a = opa_cast_string(v);
         double n;
-
+        // NOTE: we're only using opa_atof64 for validation here
         if (opa_atof64(a->v, a->len, &n) != 0)
         {
             return NULL;
         }
 
-        return opa_number_float(n);
+        return opa_number_ref(a->v, a->len);
     }
     default:
         return NULL;