This is the bare minimum to address #3298.

Little change in the benchmarks, there's much potential in improving our
number handling, but that's something we should do purposefully later.

```
name                                    old time/op  new time/op  delta
WASMArrayIteration/10-16                 420µs ± 2%   404µs ± 2%   -3.82%  (p=0.008 n=5+5)
WASMArrayIteration/100-16                435µs ± 6%   410µs ± 4%     ~     (p=0.056 n=5+5)
WASMArrayIteration/1000-16               582µs ± 2%   554µs ± 4%   -4.86%  (p=0.016 n=5+5)
WASMArrayIteration/10000-16             2.22ms ± 4%  2.23ms ± 6%     ~     (p=0.690 n=5+5)
WASMSetIteration/10-16                   426µs ± 6%   401µs ± 4%   -5.84%  (p=0.032 n=5+5)
WASMSetIteration/100-16                  442µs ± 3%   427µs ± 3%   -3.45%  (p=0.032 n=5+5)
WASMSetIteration/1000-16                 659µs ± 2%   673µs ± 2%     ~     (p=0.151 n=5+5)
WASMSetIteration/10000-16               3.86ms ± 7%  3.77ms ± 5%     ~     (p=0.310 n=5+5)
WASMObjectIteration/10-16                428µs ± 3%   426µs ± 5%     ~     (p=1.000 n=5+5)
WASMObjectIteration/100-16               433µs ± 6%   447µs ± 4%     ~     (p=0.222 n=5+5)
WASMObjectIteration/1000-16              687µs ± 3%   702µs ± 3%     ~     (p=0.151 n=5+5)
WASMObjectIteration/10000-16            3.88ms ± 3%  4.13ms ± 6%   +6.42%  (p=0.032 n=5+5)
WASMLargeJSON/10x10-16                   461µs ± 5%   448µs ± 4%     ~     (p=0.421 n=5+5)
WASMLargeJSON/10x100-16                  484µs ± 3%   497µs ± 9%     ~     (p=0.548 n=5+5)
WASMLargeJSON/10x1000-16                 831µs ± 4%   917µs ± 2%  +10.29%  (p=0.008 n=5+5)
WASMLargeJSON/10x10000-16               5.63ms ± 3%  6.12ms ± 5%   +8.67%  (p=0.008 n=5+5)
WASMLargeJSON/100x100-16                 688µs ± 2%   730µs ± 3%   +6.17%  (p=0.008 n=5+5)
WASMLargeJSON/100x1000-16               3.20ms ±10%  3.27ms ± 5%     ~     (p=0.548 n=5+5)
WASMVirtualDocs/total=1/hit=1-16         421µs ± 2%   410µs ± 3%     ~     (p=0.095 n=5+5)
WASMVirtualDocs/total=10/hit=1-16        422µs ± 3%   406µs ± 5%     ~     (p=0.095 n=5+5)
WASMVirtualDocs/total=100/hit=1-16       430µs ± 5%   420µs ± 4%     ~     (p=0.548 n=5+5)
WASMVirtualDocs/total=1000/hit=1-16      460µs ± 2%   457µs ± 5%     ~     (p=0.841 n=5+5)
WASMVirtualDocs/total=10/hit=10-16       413µs ± 6%   429µs ± 5%     ~     (p=0.056 n=5+5)
WASMVirtualDocs/total=100/hit=10-16      417µs ± 6%   436µs ± 2%     ~     (p=0.056 n=5+5)
WASMVirtualDocs/total=1000/hit=10-16     473µs ± 5%   480µs ± 2%     ~     (p=0.690 n=5+5)
WASMVirtualDocs/total=100/hit=100-16     453µs ± 2%   454µs ± 8%     ~     (p=0.421 n=5+5)
WASMVirtualDocs/total=1000/hit=100-16    503µs ± 4%   499µs ± 6%     ~     (p=1.000 n=5+5)
WASMVirtualDocs/total=1000/hit=1000-16   843µs ± 1%   823µs ± 5%     ~     (p=0.222 n=5+5)
```

------
```
$ opa eval -t wasm --format=pretty 'to_number("1234567890") == 1234567890'
true
$ opa eval -t wasm --format=pretty 'to_number("1234567890")'
1234567890
```